Work on volume scaling

+++ Emulator/SID/SIDTypes.h
@@ -103,12 +103,12 @@ typedef struct
     SamplingMethod sampling;
     
     // Input channel volumes and pan settings
-    double vol[4];
-    double pan[4];
+    long vol[4];
+    long pan[4];
     
     // Output channel volumes
-    double volL;
-    double volR;
+    long volL;
+    long volR;
 
 }
 SIDConfig;
